+++ src/rebound.h
@@ -682,7 +682,7 @@ DLLEXPORT int reb_simulation_output_screenshot(struct reb_simulation* r, const c
 // Timestepping
 
 // Advance simulation by N_steps timesteps.
-DLLEXPORT void reb_simulation_steps(struct reb_simulation* const r, unsigned int N_steps);
+DLLEXPORT void reb_simulation_steps(struct reb_simulation* const r, size_t N_steps);
 // Integrate simulation to at least time tmax (see exact_finish_time).
 DLLEXPORT enum REB_STATUS reb_simulation_integrate(struct reb_simulation* const r, double tmax);
 // Synchronize simulation if safe_mode is turned off by integrator to get physical coordinates.
